Add Gift Profile Email Button
Added a button on the Gift Profile tab of the Recipients page so that if
they haven't filled out their favorites yet the User can re-send the email.
This will be useful if they accidentally get the email address wrong or
decided not to send the email in the initial recipient creation. It pops
up a little toast message in the corner to let them know the status of
how the button worked out.