Remove Unused Code
Getting rid of props and code that we're not using in RecipientProfile and
GeneralInfo that's not being used anywhere else. Also added back in the
very first "import React" line because I think that we need to keep those
around for React components to make it clear what the files are at a glance.