Show Recipient Update Errors
Adding a snack bar to the EditRecipient component that shows the user any
errors captured from submitting the form. This is so that the user knows
they need to fix something in order to save the form. The most common error
is having an invalid email address which causes a message to be sent back
from GraphQL, so the format is a bit different and we have to take that into
account when grabbing the message. Also getting rid of "phone" because we're
not using it or showing it anywhere yet. Also adding default empty strings
to fields in case they're null, because the console yells at nulls in the
input value fields, and then adding some space between the "Save" and "Cancel"
buttons.