Send Favorites Email to Recipient Upload
Updated the "Import a List" for the Recipients so that after we create all
of the recipients in the csv, we capture their ids and can then pass them
into the success modal to be able to send all of them the favorites survey.
We also had to save the email addresses from the csv file because apparently
we weren't doing that before. Oops.