Add Profile Update Error Handling
Added a little snack bar to the EditProfile page so that if there's a
problem when the user is trying to update their profile it will let them
know. The most common problem is having a bad email address or phone
number since those are special AWS types that will throw errors if they
don't validate in the backend. We can also catch a normal error about the
name being missing, so made sure the catch can handle both formats.