Fix NavLink
When we updated the package-lock.json, it updated react-router-dom and
there was a slightly different thing going on where we needed to make sure
to add the "end" prop to the NavLink so that the Dashboard menu item wasn't
active on every single page because it's technically the parent route of
all the other pages. So...fixed that.