Use holidaysData for Upcoming Gifts
Using an array reduce to go through all of the holidaysData to grab the
next three holidays. It only really works if the holidaysData is in order.
Also updating the holidaysData with more accurate data for this year and
early next. It's not a super efficient algorithm or anything, but it gets
the job done.