Allow ProfileFavorite Read
Turns out that the Account Users couldn't access the ProfileFavorites on
their recipients to see their favorites because they were created by the IAM
user in the lambda function. Updating it so that all of the ProfileFavorites
are readable by logged in users as a short-term solution to allow users
to see their recipient's favorites. We'll need to make it more secure and
update it later so that the ProfileFavorites have a Company-Level accessGroup
that the USers are in.