Set Start Dates a Week Out
We only want to allow users to select dates that are at least a week in
the future to allow us to get their gifts ready to ship them and stuff.
Making it so that the minimum date they can select in the calendars is
a week out, and the default date before they've picked anything is also
a week out.