Sort Recipients
Sorting the recipients alphabetically so that it's easier to scroll down
and find someone without using the search bar. Updating both the main
Recipients page and the Choose a Recipient list, but because it's using
the Array sort method, it does move it in place and both places get updated
when the other one runs. I think.