Add no_shipping to Gifts
Adding in a no_shipping field to Gifts to indicate that they do not need
to be shipped. This applies to Digital Gift Cards and other services
that can be delivered immediately to recipients. These will affect the way
that we handle scheduling and checkout shipping price on the front-end.