Support No Shipping Fee and Instant Shipping
Added in support for when Gifts have no_shipping_fee and/or can_schedule_immediately.
This lets users schedule the gift immediately if they want to (instead of
a week out) and removes the shipping fee from the total price and the
shipping fee selection box on the Checkout page.