Update Recipient Email
Our mail-tester.com score was 7.3/10, so followed the instructions to fix
some of the problems to increase our score to 9/10. Had to add a DMARC
record to the DNS, add a List-Unsubscribe header, add a plain-text version,
and remove some broken links (the links had been copied from the first
email template made for the Cognito sign-up. Should have verified them...).
To add the header, had to move to using the Raw Email command and brought
in the mailcomposer library to handle all of that stuff.