Update Recipient Subject Line
Updating the subject line for the Recipients Favorite's Profile Email so
that it will use the Company Name for the recipient if they have one attached
to their Recipient Object, otherwise it uses the word "Someone" because
recipients can either be employees OR clients, so it's not necessarily
from their Employer/Company. Also updating the "from" in the email to
say it's from "Zip Zap Gifts" because otherwise it will show up in inboxes
as from "noreply" which doesn't look great.