Use Recipient Data for Dashboard Calendar
Updating the Dashboard Calendar to use Recipient Data. Getting all of their
birthdays and anniversaries and setting the year to this year so that they'll
show up on the calendar. There's a possibility that it could have performance
problems if they had a ton of recipients, but we'll cross that bridge when
we get to it.