Add Calendar Event Hover
Adding a Tooltip to the giftDates on the Dashboard Calendar so that you can
hover over them and see what exactly they are. Had to get rid of the neat
date hover-over effect because it was making it so that the tooltip wouldn't
work...maybe we could bring it back someday...also had to get rid of the
absolute positioning because all of the events were directly on top of each
other. Having them line up in a row is better for now....we can look into
solutions for when there are many many people's events on a single day when
we get clients that have that problem.