Remove Touch Delay
Setting enterTouchDelay={0} for the calendar event tool tip so that it pops up immediately on a touch in mobile. It is unintuitive for users to do a long press to see the name of the event on the calendar, so making it easier for them to easily see whose birthdays and anniversaries are coming up on their phones.