Add Recipient Group to Creation
Adding in Group to Recipient creation, including the csv import and single
add form. Replacing "Department" with "Group" in both cases since we're
not using Department right now. This includes the csv upload template
and example template table. In the template, also updated "Title" to
"Job Title" so that it matches all of the other places. It would be bad if
their job titles didn't get uploaded because they were using the online
example instead of the csv template...so, making them match.