Create Company with User
Making it so that we create the User's Company when we first add the User
to GraphQL. This is so that all of the Recipients they create later will
be able to have a CompanyID even if the User hasn't gone and given their
company a proper name yet. This will allow us to use the Company Name
in Recipient emails and stuff later when the User fills it in so we don't
have to hook up anything later.