Include Recipient Company ID
Adding the Company ID to the Recipient whenever they're created so that
we can use the Company name in emails and stuff later. If the company doesn't
have a name yet, that's ok because when the User finally adds a name, all
of the Recipients will already be hooked up. Not gonna lie, I don't 100%
trust this code...I don't like having to depend on the userData being there
when I feel like there's a possibility that it could not be there. But it's
kinda late, and I'm tired, and both of my tests worked, so...it's good enough
for now...I guess user should probably be an empty object and the null dot
operator should make it so that it doesn't crash or anything...we'll see...